Move CheckError from remote to transport
We'd like to use this inside transport, but we can't without introducing
a dependency cycle. Moving CheckError into transport for now to fix that
until we have a better place for it.
